Фиксирует промежуточное состояние
