+++ src/utils/render.js
@@ -1,4 +1,5 @@
 import Abstract from "../view/abstract.js";
+import {getRandomInteger} from "./common.js";
 
 export const RenderPosition = {
   AFTERBEGIN: `afterbegin`,
@@ -68,4 +69,50 @@ export const remove = (component) => {
 
   component.getElement().remove();
   component.removeElement();
+};
+
+export const generateDescription = () => {
+  const DESCRIPTIONS = [
+    `Lorem ipsum dolor sit amet, consectetur adipiscing elit.`,
+    `Cras aliquet varius magna, non porta ligula feugiat eget.`,
+    `Fusce tristique felis at fermentum pharetra.`,
+    `Aliquam id orci ut lectus varius viverra.`,
+    `Nullam nunc ex, convallis sed finibus eget, sollicitudin eget ante.`,
+    `Phasellus eros mauris, condimentum sed nibh vitae, sodales efficitur ipsum.`,
+    `Sed blandit, eros vel aliquam faucibus, purus ex euismod diam, eu luctus nunc ante ut dui.`,
+    `Sed sed nisi sed augue convallis suscipit in sed felis.`,
+    `Aliquam erat volutpat.`,
+    `Nunc fermentum tortor ac porta dapibus.`,
+    `In rutrum ac purus sit amet tempus.`
+  ];
+
+  const DESCRIPTIONS_NUMBER_MIN = 1;
+  const DESCRIPTIONS_NUMBER_MAX = 5;
+
+  const descriptions = [];
+  let temp = DESCRIPTIONS.slice();
+  let size = getRandomInteger(DESCRIPTIONS_NUMBER_MIN, DESCRIPTIONS_NUMBER_MAX);
+
+  for (let i = 1; i <= size; i++) {
+    const index = getRandomInteger(0, temp.length - 1);
+    descriptions.push(temp[index]);
+    temp.splice(index, 1);
+  }
+
+  return descriptions;
+};
+
+export const generatePhotos = () => {
+  const PHOTOS_NUMBER_MIN = 1;
+  const PHOTOS_NUMBER_MAX = 5;
+
+  const photos = [];
+
+  let size = getRandomInteger(PHOTOS_NUMBER_MIN, PHOTOS_NUMBER_MAX);
+
+  for (let i = 1; i <= size; i++) {
+    photos.push(`http://picsum.photos/248/152?r=${i}`);
+  }
+
+  return photos;
 };
